requirements.txt
colour==0.1.5
django-import-export==2.9.0
dj-database-url==0.5.0
Django==4.0.3
#mysqlclient==2.1.0
mysql-connector-python==8.0.28
pandas
djangorestframework==3.13.1